add browser screenshot support
working toward #2265
